# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Chat application API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /groups:
    post:
      operationId: createGroup
      tags: [groups]
      summary: Create a group conversation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateGroupInput"
      responses:
        "201":
          description: Created group conversation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Conversation"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /groups/{id}:
    patch:
      operationId: updateGroup
      tags: [groups]
      summary: Update a group's title or icon (creator or admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateGroupInput"
      responses:
        "200":
          description: Updated group conversation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Conversation"
  /groups/{id}/members:
    post:
      operationId: addGroupMember
      tags: [groups]
      summary: Add a member to a group (creator or admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddGroupMemberInput"
      responses:
        "200":
          description: Updated group conversation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Conversation"
  /groups/{id}/members/{userId}:
    delete:
      operationId: removeGroupMember
      tags: [groups]
      summary: Remove a member from a group (creator, admin, or self-leave)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated group conversation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Conversation"
...
  /e2ee/keys:
    post:
      operationId: publishE2eeKeys
      tags: [e2ee]
      summary: Publish this user's public key bundle (identity, signed prekey, one-time prekeys)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/E2eePublishKeysInput"
      responses:
        "200":
          description: Key status after publish
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/E2eeKeyStatus"
    get:
      operationId: getE2eeKeyStatus
      tags: [e2ee]
      summary: Get this user's published key status and remaining one-time prekey count
      responses:
        "200":
          description: Key status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/E2eeKeyStatus"
  /e2ee/bundles:
    get:
      operationId: getE2eeBundles
      tags: [e2ee]
      summary: Fetch prekey bundles for users (consumes one one-time prekey per user)
      parameters:
        - name: userIds
          in: query
          required: true
          description: Comma-separated user ids
          schema:
            type: string
      responses:
        "200":
          description: Bundles for users that have published keys
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/E2eeBundle"
...
  /admin/sla-settings:
    put:
      operationId: updateSlaSettings
      tags: [sla]
      summary: Update SLA time limits (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SlaSettingsInput"
              x-body-name: UpdateSlaSettingsBody
      responses:
        "200":
          description: Updated settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SlaSettings"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    User:
      type: object
      required: [id, name, email, role, isActive, isOnline]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [admin, agent, user]
        empCode:
          type: ["string", "null"]
        mobile:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        isActive:
          type: boolean
        isOnline:
          type: boolean
        mustChangePassword:
          type: boolean
        lastSeenAt:
          type: ["string", "null"]
        createdAt:
          type: string
...
    CreateGroupInput:
      type: object
      required: [title, memberIds]
      properties:
        title:
          type: string
          minLength: 1
        memberIds:
          type: array
          items:
            type: integer
        iconUrl:
          type: string
    UpdateGroupInput:
      type: object
      properties:
        title:
          type: string
          minLength: 1
        iconUrl:
          type: ["string", "null"]
    AddGroupMemberInput:
      type: object
      required: [userId]
      properties:
        userId:
          type: integer
    Conversation:
      type: object
      required: [id, otherUser, unreadCount, updatedAt, mode, aiResponseCount, language, adminEscalated, type]
      properties:
        id:
          type: integer
        type:
          type: string
          enum: [caller, direct, group]
        title:
          type: ["string", "null"]
        iconUrl:
          type: ["string", "null"]
        createdById:
          type: ["integer", "null"]
        members:
          type: ["array", "null"]
          items:
            $ref: "#/components/schemas/User"
        otherUser:
          oneOf:
            - $ref: "#/components/schemas/User"
            - type: "null"
        lastMessage:
          oneOf:
            - $ref: "#/components/schemas/Message"
            - type: "null"
        unreadCount:
          type: integer
        updatedAt:
          type: string
        mode:
          type: string
          enum: [ai, human]
        selectedCategoryId:
          type: ["integer", "null"]
        escalationReason:
          type: ["string", "null"]
        aiResponseCount:
          type: integer
        language:
          type: string
        adminEscalated:
          type: boolean
        slaStatus:
          type: string
          enum: [open, resolved]
        openedAt:
          type: ["string", "null"]
        firstResponseAt:
          type: ["string", "null"]
        awaitingReplySince:
          type: ["string", "null"]
        resolvedAt:
          type: ["string", "null"]
        archivedAt:
          type: ["string", "null"]
...
    Message:
      type: object
      required: [id, conversationId, senderId, isAi, status, createdAt]
      properties:
        id:
          type: integer
        conversationId:
          type: integer
        senderId:
          type: ["integer", "null"]
        senderName:
          type: ["string", "null"]
        isAi:
          type: boolean
        content:
          type: ["string", "null"]
        contentEn:
          type: ["string", "null"]
        encrypted:
          type: boolean
        envelope:
          oneOf:
            - $ref: "#/components/schemas/MessageEnvelopeOut"
            - type: "null"
        attachmentUrl:
          type: ["string", "null"]
        attachmentType:
          type: ["string", "null"]
        attachmentName:
          type: ["string", "null"]
        status:
          type: string
          enum: [sent, delivered, read]
        createdAt:
          type: string
    MessageEnvelopeOut:
      type: object
      required: [type, body]
      properties:
        type:
          type: integer
        body:
          type: string
    E2eePublishKeysInput:
      type: object
      required:
        [registrationId, identityKey, signedPrekeyId, signedPrekeyPub, signedPrekeySignature]
      properties:
        registrationId:
          type: integer
        identityKey:
          type: string
        signedPrekeyId:
          type: integer
        signedPrekeyPub:
          type: string
        signedPrekeySignature:
          type: string
        oneTimePrekeys:
          type: array
          items:
            $ref: "#/components/schemas/E2eeOneTimePrekeyInput"
    E2eeOneTimePrekeyInput:
      type: object
      required: [keyId, pubKey]
      properties:
        keyId:
          type: integer
        pubKey:
          type: string
    E2eeKeyStatus:
      type: object
      required: [published, oneTimePrekeyCount]
      properties:
        published:
          type: boolean
        registrationId:
          type: ["integer", "null"]
        identityKey:
          type: ["string", "null"]
        oneTimePrekeyCount:
          type: integer
    E2eeBundle:
      type: object
      required:
        [userId, registrationId, identityKey, signedPrekeyId, signedPrekeyPub, signedPrekeySignature]
      properties:
        userId:
          type: integer
        registrationId:
          type: integer
        identityKey:
          type: string
        signedPrekeyId:
          type: integer
        signedPrekeyPub:
          type: string
        signedPrekeySignature:
          type: string
        preKeyId:
          type: ["integer", "null"]
        preKeyPub:
          type: ["string", "null"]
...
    SlaSettings:
      type: object
      required: [firstResponseMins, waitingReplyMins, resolutionHours, updatedAt]
      properties:
        firstResponseMins:
          type: integer
        waitingReplyMins:
          type: integer
        resolutionHours:
          type: integer
        updatedAt:
          type: string
    SlaSettingsInput:
      type: object
      properties:
        firstResponseMins:
          type: integer
          minimum: 1
        waitingReplyMins:
          type: integer
          minimum: 1
        resolutionHours:
          type: integer
          minimum: 1
...